+++ api/src/opentrons/drivers/mag_deck/__init__.py
@@ -1,5 +1,7 @@
-from opentrons.drivers.mag_deck.driver import MagDeck
+from opentrons.drivers.mag_deck.driver import MagDeck, SimulatingDriver
+
 
 __all__ = [
-    'MagDeck'
+    'MagDeck',
+    'SimulatingDriver'
 ]
