+++ api/src/opentrons/drivers/temp_deck/__init__.py
@@ -1,5 +1,6 @@
-from opentrons.drivers.temp_deck.driver import TempDeck
+from opentrons.drivers.temp_deck.driver import TempDeck, SimulatingDriver
 
 __all__ = [
-    'TempDeck'
+    'TempDeck',
+    'SimulatingDriver'
 ]
