+++ api/src/opentrons/hardware_control/modules/thermocycler.py
@@ -1,116 +1,12 @@
 import asyncio
 from typing import Union, Optional, List, Callable
 from opentrons.drivers.thermocycler.driver import (
-    Thermocycler as ThermocyclerDriver)
+    SimulatingDriver, Thermocycler as ThermocyclerDriver)
 import logging
 from ..execution_manager import ExecutionManager
 from . import types, update, mod_abc
 
 MODULE_LOG = logging.getLogger(__name__)
-
-
-class SimulatingDriver:
-    def __init__(self):
-        self._target_temp = None
-        self._ramp_rate = None
-        self._hold_time = None
-        self._active = False
-        self._port = None
-        self._lid_status = 'open'
-        self._lid_target = None
-        self._lid_heating_active = False
-
-    async def open(self):
-        self._lid_status = 'open'
-        return self._lid_status
-
-    async def close(self):
-        self._lid_status = 'closed'
-        return self._lid_status
-
-    @property
-    def status(self):
-        return 'holding at target' if self._active else 'idle'
-
-    @property
-    def lid_status(self):
-        return self._lid_status
-
-    @property
-    def ramp_rate(self):
-        return self._ramp_rate
-
-    @property
-    def hold_time(self):
-        # Simulating driver acts as if cycles end immediately
-        return 0
-
-    @property
-    def temperature(self):
-        return self._target_temp
-
-    @property
-    def target(self):
-        return self._target_temp
-
-    @property
-    def lid_target(self):
-        return self._lid_target
-
-    @property
-    def lid_temp_status(self):
-        return 'holding at target' if self._lid_heating_active else 'idle'
-
-    @property
-    def lid_temp(self):
-        return self._lid_target
-
-    async def connect(self, port):
-        self._port = port
-
-    def disconnect(self):
-        self._port = None
-
-    async def set_temperature(self,
-                              temp: float,
-                              hold_time: float = None,
-                              ramp_rate: float = None,
-                              volume: float = None) -> None:
-        self._target_temp = temp
-        self._hold_time = hold_time
-        self._ramp_rate = ramp_rate
-        self._active = True
-
-    async def set_lid_temperature(self, temp: Optional[float]):
-        """ Set the lid temperature in deg Celsius """
-        self._lid_heating_active = True
-        self._lid_target = temp
-
-    async def deactivate_lid(self):
-        self._lid_heating_active = False
-        self._lid_target = None
-
-    async def deactivate_block(self):
-        self._target_temp = None
-        self._ramp_rate = None
-        self._hold_time = None
-        self._active = None
-
-    async def deactivate_all(self):
-        self._target_temp = None
-        self._ramp_rate = None
-        self._hold_time = None
-        self._active = None
-        self._lid_heating_active = False
-        self._lid_target = None
-
-    async def get_device_info(self):
-        return {'serial': 'dummySerialTC',
-                'model': 'dummyModelTC',
-                'version': 'dummyVersionTC'}
-
-    async def enter_programming_mode(self):
-        pass
 
 
 class Thermocycler(mod_abc.AbstractModule):
@@ -123,7 +19,8 @@ class Thermocycler(mod_abc.AbstractModule):
                     execution_manager: ExecutionManager,
                     interrupt_callback: types.InterruptCallback = None,
                     simulating: bool = False,
-                    loop: asyncio.AbstractEventLoop = None):
+                    loop: asyncio.AbstractEventLoop = None,
+                    sim_model: str = None):
         """Build and connect to a Thermocycler
         """
 
@@ -131,7 +28,8 @@ class Thermocycler(mod_abc.AbstractModule):
                   interrupt_callback=interrupt_callback,
                   simulating=simulating,
                   loop=loop,
-                  execution_manager=execution_manager)
+                  execution_manager=execution_manager,
+                  sim_model=sim_model)
         await mod._connect()
         return mod
 
@@ -149,10 +47,11 @@ class Thermocycler(mod_abc.AbstractModule):
     @staticmethod
     def _build_driver(
             simulating: bool,
+            sim_model: str,
             interrupt_cb: Callable[[str], None] = None)\
             -> Union['SimulatingDriver', 'ThermocyclerDriver']:
         if simulating:
-            return SimulatingDriver()
+            return SimulatingDriver(sim_model=sim_model)
         else:
             return ThermocyclerDriver(interrupt_cb)
 
@@ -168,7 +67,10 @@ class Thermocycler(mod_abc.AbstractModule):
                          loop=loop,
                          execution_manager=execution_manager)
         self._interrupt_cb = interrupt_callback
-        self._driver = self._build_driver(simulating, interrupt_callback)
+        self._driver = self._build_driver(
+            simulating,
+            sim_model or self.model(),
+            interrupt_callback)
 
         self._total_cycle_count: Optional[int] = None
         self._current_cycle_index: Optional[int] = None
